src/use-cases/authenticate/authenticate.ts
import { compare } from "bcryptjs";
import { UserModel } from "../../models/user-model";
import { UserRepository } from "../../repositories/user-repository";
import { InvalidCredentialsError } from "../errors/invalid-credentials-error";

interface AuthenticateUseCaseRequest {
  email: string;
  password: string;
}

interface AuthenticateUseCaseResponse {
  user: UserModel;
}

export class AuthenticateUseCase {
  constructor(private userRepository: UserRepository) {}

  async exec({
    email,
    password,
  }: AuthenticateUseCaseRequest): Promise<AuthenticateUseCaseResponse> {
    const user = await this.userRepository.findByEmail(email);

    if (!user) {
      throw new InvalidCredentialsError();
    }

    const doesPasswordMatches = await compare(password, user.password);

    if (!doesPasswordMatches) {
      throw new InvalidCredentialsError();
    }

    return { user };
  }
}
